requirements.txt
absl-py==0.9.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.12
argh==0.26.2
asn1crypto==1.3.0
astor==0.8.0
astroid==2.3.3
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
backports.functools-lru-cache==1.6.1
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.1.7
beautifulsoup4==4.9.0
bert-serving-client==1.10.0
bert-serving-server==1.10.0
bleach==3.1.4
boto==2.49.0
boto3==1.9.66
botocore==1.12.189
cachetools==4.1.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.4.1
clyent==1.2.2
colorama==0.4.3
conda==4.8.3
conda-build==3.18.9
conda-package-handling==1.6.0
conda-verify==3.4.2
cryptography==2.8
cycler==0.10.0
cymem==1.31.2
cytoolz==0.9.0.1
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
dill==0.2.9
dm-tree==0.1.5
docutils==0.16
en-core-web-lg==2.0.0
en-core-web-md==2.0.0
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
flake8==3.7.9
flashtext==2.7
future==0.18.2
gast==0.3.3
gensim==3.8.0
glob2==0.7
google-api-core==1.17.0
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-storage==1.28.0
google-pasta==0.2.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
GPUtil==1.4.0
graphviz==0.8.4
grpcio==1.24.3
h5py==2.10.0
helpdev==0.6.10
idna==2.9
imagesize==1.2.0
importlib-metadata==1.5.0
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jdcal==1.4.1
jedi==0.14.1
Jinja2==2.11.2
jmespath==0.9.4
joblib==0.14.1
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==1.2.6
jupyterlab-server==1.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==21.1.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
libarchive-c==2.9
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
menuinst==1.4.16
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mock==4.0.2
msgpack==0.5.6
msgpack-numpy==0.4.4.3
murmurhash==0.28.0
mxnet-cu92==1.5.0
navigator-updater==0.2.1
nbconvert==5.6.1
nbformat==5.0.4
nltk==3.4.5
notebook==6.0.3
numpy==1.17.5
numpydoc==0.9.2
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.3
opt-einsum==3.2.1
packaging==20.3
pandas==1.0.1
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pkginfo==1.5.0.1
plac==0.9.6
pluggy==0.13.1
portpicker==1.3.1
preshed==1.0.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.4
psutil==5.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.20
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.6.1
pylint==2.4.4
pymssql==2.1.4
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==12.7.1
PyQtWebEngine==5.12.1
pyreadline==2.1
pyrsistent==0.16.0
PySocks==1.7.1
pyspellchecker==0.5.4
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.8
pytz==2020.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==18.1.1
QDarkStyle==2.8.1
QtAwesome==0.7.0
qtconsole==4.7.3
QtPy==1.9.0
regex==2018.7.11
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rope==0.16.0
rsa==4.0
Rtree==0.9.4
ruamel-yaml==0.15.87
s3transfer==0.1.13
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
smart-open==1.11.1
snowballstemmer==2.0.0
sortedcontainers==2.1.0
soupsieve==2.0
spacy==2.0.12
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.0.1
spyder-kernels==1.8.1
tensorboard==2.1.1
tensorflow==1.13.1
tensorflow-addons==0.8.3
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==6.10.3
toolz==0.10.0
torch==1.3.1
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
typed-ast==1.4.1
typeguard==2.7.1
ujson==2.0.3
urllib3==1.25.8
virtualenv==16.7.5
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.10.11
xlrd==1.2.0
xmltodict==0.12.0
yapf==0.29.0
zipp==3.1.0
